Fork off the `get-azure-pipelines-artifact` Action
Much of what the `setup-git-for-windows-sdk` Action does is to download
an artifact (in a robust fashion, retrying a couple times if necessary)
and caching it.

This is useful for other things, too.

Let's transmogrify the Action that is rather specific for Git for
Windows' use case to allow downloading arbitrary artifacts from
arbitrary Azure Pipelines.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>