Optionally override the "strip prefix"
Typically, build artifacts are downloaded in the form of a `.zip` file
that contains the files, where the original path is prefixed by the
artifact name. For example, an artifact name "heya" will be reflected in
the downloaded `.zip` file by having all files inside the `heya/`
subdirectory. For that reason, this Action strips that prefix from the
extracted files.

Under certain circumstances, it may be desirable to strip an even longer
prefix, e.g. when downloading Git's `vcpkg` artifact. The corresponding
files are all inside the `compat/vcbuild/vcpkg/` subdirectory, and if we
want to cache the artifact files, we can only download them to an empty
(or non-existing) directory, and `compat/` is definitely not empty once
Git's source code is checked out.

So let's offer the option to override the strip prefix.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>