When extracting into a non-empty directory, cache the `.zip` file
Since the `@actions/cache` functionality has _no_ option to include only
a given set of files (unless that very set of files is known _when
restoring the cache_, not only when saving it), we have no chance of
caching files directly when we write them into non-empty directories.

To work around that, let's special-case non-empty directories, caching
the downloaded `.zip` file instead (and obviously paying the slight
performance penalty of `unzip`ping it in that scenario even when
restoring the cache, which is slower than extracting a `.tgz`).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>