test(downloader): prepare for an upcoming change
We are about to _always_ query the artifact metadata, even when not
downloading the artifact but restoring it from the cache.

Therefore, we need a fixture that artifact metadata. And since the logs
of the original build are gone, we'll have to use a newer one.

While at it, we already prepare the mocked `fetch()` to return the
metadata in the second call.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>