ci: be more specific which artifact to download
The `sparse` job was recently taught to generate artifacts for newer
Ubuntu versions, in preparation for GitHub Actions dropping support for
hosted Ubuntu 16.04 runners.

Therefore, the test started to fail, because it no longer had a single
artifact to choose from, and had no chance to guess which one was
intended.

Let's just specify which artifact to use.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>