ci: allow the PR build to re-run after repackaging
We just added a workflow that repackages the Action into `dist/`
whenever Dependabot updates a dependency that requires it.

However, since it is now a _workflow_ that pushes the new commit, the PR
build will not trigger _another_ workflow, and as a result the PR build
will not run.

That's not what we want.

To remedy this, we added a new environment, called
`git-for-windows-ci-push`, that contains a secret token,
`GIT_FOR_WINDOWS_CI_PUSH`, that can be used to push the new commit.
Since that token is not bound to a GitHub workflow, the PR build will be
kicked off, and all is fine and dandy again.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>