Fix race between two 'finish' events
When unzipping the file that is being downloaded, there are two streams
whose `finish` event are eagerly awaited: the stream that unpacks the
last entry of the .zip file and the stream that processes the .zip file
itself.

When this code was originally written, the stream that extracted the
.zip entry always sent its `finish` event before the stream that
processed the enclosing .zip file.

However, with our upgrade to node.js v20.x, this seems no longer to be
true. As a consequence, the code that wants to verify that all expected
bytes were extracted now _sometimes_ runs too early, and shouts
"failure" just before the code that extracts the bytes runs and all
files are written correctly.

Let's work around it by ensuring that both `finish` events were received
before determining success or failure.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>